requirements.txt
PyYAML
Flask-SQLAlchemy
Flask
Flask-login
werkzeug
SQLAlchemy
python-dotenv
requests